bigint

+++ db/bloodhound.sql
@@ -38,7 +38,7 @@ CREATE TABLE `destroyed_users` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
 
 CREATE TABLE `occurrences` (
-  `gbifID` int(11) UNSIGNED NOT NULL,
+  `gbifID` bigint(11) UNSIGNED NOT NULL,
   `occurrenceID` varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_bin DEFAULT NULL,
   `dateIdentified` text CHARACTER SET utf8mb4 COLLATE utf8mb4_bin,
   `decimalLatitude` text CHARACTER SET utf8mb4 COLLATE utf8mb4_bin,
@@ -60,12 +60,12 @@ CREATE TABLE `occurrences` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin;
 
 CREATE TABLE `occurrence_determiners` (
-  `occurrence_id` int(11) UNSIGNED NOT NULL,
+  `occurrence_id` bigint(11) UNSIGNED NOT NULL,
   `agent_id` int(11) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin;
 
 CREATE TABLE `occurrence_recorders` (
-  `occurrence_id` int(11) UNSIGNED NOT NULL,
+  `occurrence_id` bigint(11) UNSIGNED NOT NULL,
   `agent_id` int(11) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin;
 
@@ -99,7 +99,7 @@ CREATE TABLE `taxon_determiners` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin;
 
 CREATE TABLE `taxon_occurrences` (
-  `occurrence_id` int(11) UNSIGNED NOT NULL,
+  `occurrence_id` bigint(11) UNSIGNED NOT NULL,
   `taxon_id` int(11) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin;
 
@@ -132,7 +132,7 @@ CREATE TABLE `users` (
 CREATE TABLE `user_occurrences` (
   `id` int(11) NOT NULL,
   `user_id` int(11) NOT NULL,
-  `occurrence_id` int(11) NOT NULL,
+  `occurrence_id` bigint(11) NOT NULL,
   `action` varchar(100) CHARACTER SET utf8 COLLATE utf8_bin DEFAULT NULL,
   `visible` tinyint(1) NOT NULL DEFAULT '1',
   `created` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
